+++ assets/javascripts/kitten/karl/buttons/payment-button.js
@@ -4,10 +4,21 @@ import { TagButton } from 'kitten/components/buttons/tag-button'
 import { VisaIcon } from 'kitten/components/icons/visa-icon'
 import { MastercardIcon } from 'kitten/components/icons/mastercard-icon'
 
-export const KarlPaymentButton = (props) => {
+export const KarlPaymentButtonVisa = props => {
   const text = <TagButton big={ true } icon={ true } tag="span">
                  <VisaIcon />
                </TagButton>
+
+  return (
+    <RadioButton id="payment-button" text={ text } { ...props } />
+  )
+}
+
+export const KarlPaymentButtonMastercard = props => {
+  const text = <TagButton big={ true } icon={ true } tag="span">
+                 <MastercardIcon />
+               </TagButton>
+
   return (
     <RadioButton id="payment-button" text={ text } { ...props } />
   )